Allow character sequence `<>` in PDF rendering
The PDF renderer attempts to parse HTML by means of regex, which you
[cannot do](https://stackoverflow.com/a/1732454/6278). This breaks when
encountering text like `A <> B`, which it thinks is an empty HTML tag,
causing an IndexError when trying to identify whether it's a closing tag
(`if part[0] == "/"`).

This is a small workaround for this specific issue, although the bigger
problem is really the approach of parsing HTML with regex.